Added Buffer + fixed warnings

+++ Atlas-Editor/src/Panels/ContentBrowserPanel.cpp
@@ -50,7 +50,7 @@ namespace Atlas
 
 			ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0, 0, 0, 0));
 			Ref<Texture2D> icon = GetDirectoryEntryIcon(directoryEntry);
-			ImGui::ImageButton((ImTextureID)icon->GetRendererID(), { thumbnailSize, thumbnailSize }, { 0, 1 }, { 1, 0 });
+			ImGui::ImageButton((ImTextureID)(uint64_t)icon->GetRendererID(), { thumbnailSize, thumbnailSize }, { 0, 1 }, { 1, 0 });
 
 			if (ImGui::BeginDragDropSource())
 			{
